+++ game/missiongenerator/aircraft/waypoints/refuel.py
@@ -1,12 +1,17 @@
 from dcs.point import MovingPoint
 from dcs.task import RefuelingTaskAction, ControlledTask
 
+from game.theater import OffMapSpawn
 from .pydcswaypointbuilder import PydcsWaypointBuilder
 
 
 class RefuelPointBuilder(PydcsWaypointBuilder):
     def add_tasks(self, waypoint: MovingPoint) -> None:
-        refuel = ControlledTask(RefuelingTaskAction())
-        refuel.start_probability(10)
-        waypoint.add_task(refuel)
+        offmap = isinstance(self.flight.arrival, OffMapSpawn)
+        if self.flight.divert:
+            offmap |= isinstance(self.flight.divert, OffMapSpawn)
+        if not offmap:
+            refuel = ControlledTask(RefuelingTaskAction())
+            refuel.start_probability(10)
+            waypoint.add_task(refuel)
         return super().add_tasks(waypoint)