+++ changelog.md
@@ -33,6 +33,7 @@
 * **[Options]** Option to enable unlimited fuel for AI (player and non-player flights)
 * **[Mission Generator]** F-15E Strike targets are automatically added as Mission Set 1 
 * **[Mission Generator]** Set F-14's IP waypoint according to the flight-plan's ingress point
+* **[Mission Generator]** Automatically de-spawn aircraft when arrival/divert is an off-map spawn
 
 ## Fixes
 * **[Mission Generation]** Anti-ship strikes should use "group attack" in their attack-task